Rename to .ts and lint frontend alert use cases
