+++ frontend/src/domain/use_cases/alert/getSilencedAlerts.ts
@@ -1,0 +1,13 @@
+import { getSilencedAlertsFromAPI } from '../../../api/alert'
+import { setSilencedAlerts } from '../../shared_slices/Alert'
+import { setError } from '../../shared_slices/Global'
+
+export const getSilencedAlerts = () => dispatch => {
+  getSilencedAlertsFromAPI()
+    .then(silencedAlerts => {
+      dispatch(setSilencedAlerts(silencedAlerts))
+    })
+    .catch(error => {
+      dispatch(setError(error))
+    })
+}
