+++ frontend/src/domain/use_cases/alert/reactivateSilencedAlert.ts
@@ -1,8 +1,8 @@
-import { setError } from '../../shared_slices/Global'
-import { setSilencedAlerts } from '../../shared_slices/Alert'
 import { deleteSilencedAlertFromAPI } from '../../../api/alert'
+import { setSilencedAlerts } from '../../shared_slices/Alert'
+import { setError } from '../../shared_slices/Global'
 
-const reactivateSilencedAlert = id => (dispatch, getState) => {
+export const reactivateSilencedAlert = id => (dispatch, getState) => {
   const previousSilencedAlerts = getState().alert.silencedAlerts
   const previousSilencedAlertsWithReactivatedFlag = setAlertAsReactivated(previousSilencedAlerts, id)
   dispatch(setSilencedAlerts(previousSilencedAlertsWithReactivatedFlag))
@@ -15,35 +15,35 @@ const reactivateSilencedAlert = id => (dispatch, getState) => {
   deleteSilencedAlertFromAPI(id).catch(error => {
     clearTimeout(timeout)
     dispatch(setSilencedAlerts(previousSilencedAlerts))
-    console.error(error)
     dispatch(setError(error))
   })
 }
 
-function setAlertAsReactivated (previousSilencedAlerts, id) {
+function setAlertAsReactivated(previousSilencedAlerts, id) {
   return previousSilencedAlerts.reduce((acc, alert) => {
     if (alert.id === id) {
-      const validatedAlert = Object.assign({}, alert)
+      const validatedAlert = { ...alert }
       validatedAlert.isReactivated = true
 
       acc.push(validatedAlert)
+
       return acc
     }
 
     acc.push(alert)
+
     return acc
   }, [])
 }
 
-export function removeAlert (previousAlerts, id) {
+export function removeAlert(previousAlerts, id) {
   return previousAlerts.reduce((acc, alert) => {
     if (alert.id === id) {
       return acc
     }
 
     acc.push(alert)
+
     return acc
   }, [])
 }
-
-export default reactivateSilencedAlert
